vttTranslate/requirements.txt
webvtt-py
openai
langchain
